+++ lib/osm.dart
@@ -8,4 +8,6 @@ export 'src/filter.dart';
 export 'src/pbf/exception.dart';
 export 'src/pbf/file.dart';
 export 'src/pbf/header.dart';
+export 'src/xml/change.dart';
+export 'src/xml/exception.dart';
 export 'src/subset.dart';